+++ src/types/next-auth.d.ts
@@ -12,5 +12,6 @@ declare module "next-auth" {
 declare module "next-auth/jwt" {
     interface JWT {
         isAdmin?: boolean;
+        authProvider?: string;
     }
 }